re-organized my files and adjusted array limit for user search
