+++ src/App.css
@@ -35,7 +35,7 @@ body {
 
 /* HEADER */
 header {
-  background-image: url(./camille-cox-eLoG12u76dI-unsplash.jpg);
+  background-image: url(./assets/camille-cox-eLoG12u76dI-unsplash.jpg);
   background-size: cover;
   background-repeat: no-repeat;
   height: 600px;
